soc/intel/common: Update cpu_apic_info_type struct
The patch updates total cpu count variable and total P-core count in
cpu_apic_info_type structure to `unsigned short int` to address more
cores.

TEST=Verify the build on Rex

Reviewed-on: https://review.coreboot.org/c/coreboot/+/74305
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/intel/common/block/acpi/cpu_hybrid.c
@@ -28,14 +28,14 @@ struct cpu_apic_info_type {
 	int32_t apic_ids[CONFIG_MAX_CPUS];
 
 	/* Total CPU count */
-	uint8_t total_cpu_cnt;
+	uint16_t total_cpu_cnt;
 
 	/*
 	 * Total Performance core count. This will be used
 	 * to identify the start of Efficient Cores's
 	 * APIC ID list
 	 */
-	uint8_t perf_cpu_cnt;
+	uint16_t perf_cpu_cnt;
 };
 
 static struct cpu_apic_info_type cpu_apic_info;
